separate features for SHARP display/ i2c_puppet
