+++ platforms/allwinner-d1/src/drivers/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "i2c_puppet")]
+pub mod i2c_puppet;
+#[cfg(feature = "sharp-display")]
 pub mod sharp_display;
 pub mod spim;
 pub mod twi;
